leaveroom

+++ netyap/chat/views.py
@@ -15,25 +15,25 @@ def joinChatroom(request,chatroom_id):
 
     context = RequestContext(request)
     session = request.session
-    user_id = session.get('userId')
-    #theChatRoom = Chatroom.objects.all().filter(pk = chatroom_id)[0]
+    user_id = session.get('username')
+    theChatRoom = Chatroom.objects.all().filter(chatroom_id=chatroom_id)[0]
     #chat_data = {'room_title': theChatRoom.title, 'course_name': theChatRoom.course_id, 'user_name': session.get('name'), 'instructor': theChatRoom.instructor_name}
     subscribe= SubscriberTable(
-                chatroom_id=chatroom_id,
+                chatroom_id=theChatRoom,
                 user_id=user_id,
             )
     subscribe.save()
     chat_data={}
-    chat_data['data']=Chat.objects.filter(chatroom_id)
+    chat_data['data']=Chat.objects.filter(chatroom_id=chatroom_id)
     return render_to_response('chat/chatroom.html',chat_data, context)
 
 def leaveroom(request,chatroom_id):
 
     context = RequestContext(request)
     session = request.session
-    user_id = session.get('userId')
-    instance = SubscriberTable.objects.get(chatroom_id=chatroom_id).get(user_id=user_id)
-    instance.delete()
+    user_id = session.get('username')
+    instance = SubscriberTable.objects.filter(chatroom_id=chatroom_id).filter(user_id = user_id).delete()
+    #instance.delete()
     return redirect('/')
